requirements.txt
Django>=2.0,<3.0
mysqlclient
mysql-connector-python
djangorestframework==3.8.2
django-phonenumber-field==5.0.0
phonenumbers
django_filter
drf-yasg